requirements.txt
Django>=4.2.0
djangorestframework>=3.15.0
channels>=4.0.0
channels-redis>=4.0.0
kafka-python>=2.0.2
redis>=5.0.0
psycopg2-binary>=2.9.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
python-decouple>=3.8
scikit-learn>=1.3.0
plotly>=5.17.0
websockets>=11.0.3
